Fix backward codec test after introducing the doc values skip index
