+++ lucene/test-framework/src/java/org/apache/lucene/tests/index/BaseDocValuesFormatTestCase.java
@@ -56,7 +56,6 @@ import org.apache.lucene.index.CheckIndex.Status.DocValuesStatus;
 import org.apache.lucene.index.CodecReader;
 import org.apache.lucene.index.DirectoryReader;
 import org.apache.lucene.index.DocValues;
-import org.apache.lucene.index.DocValuesSkipper;
 import org.apache.lucene.index.IndexReader;
 import org.apache.lucene.index.IndexWriter;
 import org.apache.lucene.index.IndexWriterConfig;
@@ -101,24 +100,6 @@ import org.apache.lucene.util.automaton.RegExp;
  * improved!
  */
 public abstract class BaseDocValuesFormatTestCase extends BaseIndexFileFormatTestCase {
-
-  /**
-   * Override and return {@code false} if the {@link DocValuesSkipper} produced by this format
-   * sometimes returns documents in {@link DocValuesSkipper#minDocID(int)} or {@link
-   * DocValuesSkipper#maxDocID(int)} that may not have a value.
-   */
-  protected boolean skipperHasAccurateDocBounds() {
-    return true;
-  }
-
-  /**
-   * Override and return {@code false} if the {@link DocValuesSkipper} produced by this format
-   * sometimes returns values in {@link DocValuesSkipper#minValue(int)} or {@link
-   * DocValuesSkipper#maxValue(int)} that none of the documents in the range have.
-   */
-  protected boolean skipperHasAccurateValueBounds() {
-    return true;
-  }
 
   @Override
   protected void addRandomFields(Document doc) {
@@ -908,7 +889,7 @@ public abstract class BaseDocValuesFormatTestCase extends BaseIndexFileFormatTes
     iwriter.addDocument(doc);
     doc = new Document();
     doc.add(new StringField("id", "1", Field.Store.NO));
-    doc.add(SortedDocValuesField.indexedField("field", newBytesRef("hello")));
+    doc.add(new SortedDocValuesField("field", newBytesRef("hello")));
     iwriter.addDocument(doc);
     iwriter.commit();
     iwriter.deleteDocuments(new Term("id", "1"));
@@ -919,11 +900,6 @@ public abstract class BaseDocValuesFormatTestCase extends BaseIndexFileFormatTes
 
     SortedDocValues dv = getOnlyLeafReader(ireader).getSortedDocValues("field");
     assertEquals(NO_MORE_DOCS, dv.nextDoc());
-
-    DocValuesSkipper skipper = getOnlyLeafReader(ireader).getDocValuesSkipper("field");
-    assertEquals(0, skipper.docCount());
-    skipper.advance(0);
-    assertEquals(NO_MORE_DOCS, skipper.minDocID(0));
 
     TermsEnum termsEnum = dv.termsEnum();
     assertFalse(termsEnum.seekExact(new BytesRef("lucene")));
@@ -2258,7 +2234,7 @@ public abstract class BaseDocValuesFormatTestCase extends BaseIndexFileFormatTes
     iwriter.addDocument(doc);
     doc = new Document();
     doc.add(new StringField("id", "1", Field.Store.NO));
-    doc.add(SortedSetDocValuesField.indexedField("field", newBytesRef("hello")));
+    doc.add(new SortedSetDocValuesField("field", newBytesRef("hello")));
     iwriter.addDocument(doc);
     iwriter.commit();
     iwriter.deleteDocuments(new Term("id", "1"));
@@ -2269,11 +2245,6 @@ public abstract class BaseDocValuesFormatTestCase extends BaseIndexFileFormatTes
 
     SortedSetDocValues dv = getOnlyLeafReader(ireader).getSortedSetDocValues("field");
     assertEquals(0, dv.getValueCount());
-
-    DocValuesSkipper skipper = getOnlyLeafReader(ireader).getDocValuesSkipper("field");
-    assertEquals(0, skipper.docCount());
-    skipper.advance(0);
-    assertEquals(NO_MORE_DOCS, skipper.minDocID(0));
 
     TermsEnum termsEnum = dv.termsEnum();
     assertFalse(termsEnum.seekExact(new BytesRef("lucene")));
@@ -3240,7 +3211,7 @@ public abstract class BaseDocValuesFormatTestCase extends BaseIndexFileFormatTes
     iwriter.addDocument(doc);
     doc = new Document();
     doc.add(new StringField("id", "1", Field.Store.NO));
-    doc.add(NumericDocValuesField.indexedField("field", 5));
+    doc.add(new NumericDocValuesField("field", 5));
     iwriter.addDocument(doc);
     iwriter.commit();
     iwriter.deleteDocuments(new Term("id", "1"));
@@ -3251,11 +3222,6 @@ public abstract class BaseDocValuesFormatTestCase extends BaseIndexFileFormatTes
 
     NumericDocValues dv = getOnlyLeafReader(ireader).getNumericDocValues("field");
     assertEquals(NO_MORE_DOCS, dv.nextDoc());
-
-    DocValuesSkipper skipper = getOnlyLeafReader(ireader).getDocValuesSkipper("field");
-    assertEquals(0, skipper.docCount());
-    skipper.advance(0);
-    assertEquals(NO_MORE_DOCS, skipper.minDocID(0));
 
     ireader.close();
     directory.close();
@@ -3371,7 +3337,7 @@ public abstract class BaseDocValuesFormatTestCase extends BaseIndexFileFormatTes
     iwriter.addDocument(doc);
     doc = new Document();
     doc.add(new StringField("id", "1", Field.Store.NO));
-    doc.add(SortedNumericDocValuesField.indexedField("field", 5));
+    doc.add(new SortedNumericDocValuesField("field", 5));
     iwriter.addDocument(doc);
     iwriter.commit();
     iwriter.deleteDocuments(new Term("id", "1"));
@@ -3382,11 +3348,6 @@ public abstract class BaseDocValuesFormatTestCase extends BaseIndexFileFormatTes
 
     SortedNumericDocValues dv = getOnlyLeafReader(ireader).getSortedNumericDocValues("field");
     assertEquals(NO_MORE_DOCS, dv.nextDoc());
-
-    DocValuesSkipper skipper = getOnlyLeafReader(ireader).getDocValuesSkipper("field");
-    assertEquals(0, skipper.docCount());
-    skipper.advance(0);
-    assertEquals(NO_MORE_DOCS, skipper.minDocID(0));
 
     ireader.close();
     directory.close();
@@ -3538,7 +3499,7 @@ public abstract class BaseDocValuesFormatTestCase extends BaseIndexFileFormatTes
 
     Document doc = new Document();
     doc.add(new StringField("id", "1", Field.Store.NO));
-    doc.add(SortedDocValuesField.indexedField("field", newBytesRef("hello")));
+    doc.add(new SortedDocValuesField("field", newBytesRef("hello")));
     iwriter.addDocument(doc);
     final int numEmptyDocs = atLeast(1024);
     for (int i = 0; i < numEmptyDocs; ++i) {
@@ -3553,11 +3514,6 @@ public abstract class BaseDocValuesFormatTestCase extends BaseIndexFileFormatTes
 
     SortedDocValues dv = getOnlyLeafReader(ireader).getSortedDocValues("field");
     assertEquals(NO_MORE_DOCS, dv.nextDoc());
-
-    DocValuesSkipper skipper = getOnlyLeafReader(ireader).getDocValuesSkipper("field");
-    assertEquals(0, skipper.docCount());
-    skipper.advance(0);
-    assertEquals(NO_MORE_DOCS, skipper.minDocID(0));
 
     TermsEnum termsEnum = dv.termsEnum();
     assertFalse(termsEnum.seekExact(new BytesRef("lucene")));
@@ -3578,7 +3534,7 @@ public abstract class BaseDocValuesFormatTestCase extends BaseIndexFileFormatTes
 
     Document doc = new Document();
     doc.add(new StringField("id", "1", Field.Store.NO));
-    doc.add(SortedSetDocValuesField.indexedField("field", newBytesRef("hello")));
+    doc.add(new SortedSetDocValuesField("field", newBytesRef("hello")));
     iwriter.addDocument(doc);
     final int numEmptyDocs = atLeast(1024);
     for (int i = 0; i < numEmptyDocs; ++i) {
@@ -3593,11 +3549,6 @@ public abstract class BaseDocValuesFormatTestCase extends BaseIndexFileFormatTes
 
     SortedSetDocValues dv = getOnlyLeafReader(ireader).getSortedSetDocValues("field");
     assertEquals(NO_MORE_DOCS, dv.nextDoc());
-
-    DocValuesSkipper skipper = getOnlyLeafReader(ireader).getDocValuesSkipper("field");
-    assertEquals(0, skipper.docCount());
-    skipper.advance(0);
-    assertEquals(NO_MORE_DOCS, skipper.minDocID(0));
 
     TermsEnum termsEnum = dv.termsEnum();
     assertFalse(termsEnum.seekExact(new BytesRef("lucene")));
@@ -3618,7 +3569,7 @@ public abstract class BaseDocValuesFormatTestCase extends BaseIndexFileFormatTes
 
     Document doc = new Document();
     doc.add(new StringField("id", "1", Field.Store.NO));
-    doc.add(NumericDocValuesField.indexedField("field", 42L));
+    doc.add(new NumericDocValuesField("field", 42L));
     iwriter.addDocument(doc);
     final int numEmptyDocs = atLeast(1024);
     for (int i = 0; i < numEmptyDocs; ++i) {
@@ -3633,11 +3584,6 @@ public abstract class BaseDocValuesFormatTestCase extends BaseIndexFileFormatTes
 
     NumericDocValues dv = getOnlyLeafReader(ireader).getNumericDocValues("field");
     assertEquals(NO_MORE_DOCS, dv.nextDoc());
-
-    DocValuesSkipper skipper = getOnlyLeafReader(ireader).getDocValuesSkipper("field");
-    assertEquals(0, skipper.docCount());
-    skipper.advance(0);
-    assertEquals(NO_MORE_DOCS, skipper.minDocID(0));
 
     ireader.close();
     directory.close();
@@ -3654,7 +3600,7 @@ public abstract class BaseDocValuesFormatTestCase extends BaseIndexFileFormatTes
 
     Document doc = new Document();
     doc.add(new StringField("id", "1", Field.Store.NO));
-    doc.add(SortedNumericDocValuesField.indexedField("field", 42L));
+    doc.add(new SortedNumericDocValuesField("field", 42L));
     iwriter.addDocument(doc);
     final int numEmptyDocs = atLeast(1024);
     for (int i = 0; i < numEmptyDocs; ++i) {
@@ -3669,11 +3615,6 @@ public abstract class BaseDocValuesFormatTestCase extends BaseIndexFileFormatTes
 
     SortedNumericDocValues dv = getOnlyLeafReader(ireader).getSortedNumericDocValues("field");
     assertEquals(NO_MORE_DOCS, dv.nextDoc());
-
-    DocValuesSkipper skipper = getOnlyLeafReader(ireader).getDocValuesSkipper("field");
-    assertEquals(0, skipper.docCount());
-    skipper.advance(0);
-    assertEquals(NO_MORE_DOCS, skipper.minDocID(0));
 
     ireader.close();
     directory.close();
@@ -3865,458 +3806,5 @@ public abstract class BaseDocValuesFormatTestCase extends BaseIndexFileFormatTes
 
   protected boolean codecAcceptsHugeBinaryValues(String field) {
     return true;
-  }
-
-  public void testNumericDocValuesWithSkipperSmall() throws Exception {
-    doTestNumericDocValuesWithSkipper(random().nextInt(1, 1000));
-  }
-
-  public void testNumericDocValuesWithSkipperMedium() throws Exception {
-    doTestNumericDocValuesWithSkipper(random().nextInt(1000, 20000));
-  }
-
-  @Nightly
-  public void testNumericDocValuesWithSkipperBig() throws Exception {
-    doTestNumericDocValuesWithSkipper(random().nextInt(50000, 100000));
-  }
-
-  private void doTestNumericDocValuesWithSkipper(int totalDocs) throws Exception {
-    assertDocValuesWithSkipper(
-        totalDocs,
-        new TestDocValueSkipper() {
-          @Override
-          public void populateDoc(Document doc) {
-            doc.add(NumericDocValuesField.indexedField("test", random().nextLong()));
-          }
-
-          @Override
-          public DocValuesWrapper docValuesWrapper(LeafReader leafReader) throws IOException {
-            NumericDocValues numericDocValues = leafReader.getNumericDocValues("test");
-            return new DocValuesWrapper() {
-
-              @Override
-              public int advance(int target) throws IOException {
-                return numericDocValues.advance(target);
-              }
-
-              @Override
-              public boolean advanceExact(int target) throws IOException {
-                return numericDocValues.advanceExact(target);
-              }
-
-              @Override
-              public long maxValue() throws IOException {
-                return numericDocValues.longValue();
-              }
-
-              @Override
-              public long minValue() throws IOException {
-                return numericDocValues.longValue();
-              }
-
-              @Override
-              public int docID() {
-                return numericDocValues.docID();
-              }
-            };
-          }
-
-          @Override
-          public DocValuesSkipper docValuesSkipper(LeafReader leafReader) throws IOException {
-            return leafReader.getDocValuesSkipper("test");
-          }
-        });
-  }
-
-  public void testSortedNumericDocValuesWithSkipperSmall() throws Exception {
-    doTestSortedNumericDocValuesWithSkipper(random().nextInt(1, 1000));
-  }
-
-  public void testSortedNumericDocValuesWithSkipperMedium() throws Exception {
-    doTestSortedNumericDocValuesWithSkipper(random().nextInt(1000, 20000));
-  }
-
-  @Nightly
-  public void testSortedNumericDocValuesWithSkipperBig() throws Exception {
-    doTestSortedNumericDocValuesWithSkipper(random().nextInt(50000, 100000));
-  }
-
-  private void doTestSortedNumericDocValuesWithSkipper(int totalDocs) throws Exception {
-    assertDocValuesWithSkipper(
-        totalDocs,
-        new TestDocValueSkipper() {
-          @Override
-          public void populateDoc(Document doc) {
-            for (int j = 0; j < random().nextInt(1, 5); j++) {
-              doc.add(SortedNumericDocValuesField.indexedField("test", random().nextLong()));
-            }
-          }
-
-          @Override
-          public DocValuesWrapper docValuesWrapper(LeafReader leafReader) throws IOException {
-            SortedNumericDocValues sortedNumericDocValues =
-                leafReader.getSortedNumericDocValues("test");
-            return new DocValuesWrapper() {
-              long max;
-              long min;
-
-              @Override
-              public int advance(int target) throws IOException {
-                int doc = sortedNumericDocValues.advance(target);
-                if (doc != NO_MORE_DOCS) {
-                  readValues();
-                }
-                return doc;
-              }
-
-              @Override
-              public boolean advanceExact(int target) throws IOException {
-                if (sortedNumericDocValues.advanceExact(target)) {
-                  readValues();
-                  return true;
-                }
-                return false;
-              }
-
-              private void readValues() throws IOException {
-                max = Long.MIN_VALUE;
-                min = Long.MAX_VALUE;
-                for (int i = 0; i < sortedNumericDocValues.docValueCount(); i++) {
-                  long value = sortedNumericDocValues.nextValue();
-                  max = Math.max(max, value);
-                  min = Math.min(min, value);
-                }
-              }
-
-              @Override
-              public long maxValue() {
-                return max;
-              }
-
-              @Override
-              public long minValue() {
-                return min;
-              }
-
-              @Override
-              public int docID() {
-                return sortedNumericDocValues.docID();
-              }
-            };
-          }
-
-          @Override
-          public DocValuesSkipper docValuesSkipper(LeafReader leafReader) throws IOException {
-            return leafReader.getDocValuesSkipper("test");
-          }
-        });
-  }
-
-  public void testSortedDocValuesWithSkipperSmall() throws Exception {
-    doTestSortedDocValuesWithSkipper(random().nextInt(1, 1000));
-  }
-
-  public void testSortedDocValuesWithSkipperMedium() throws Exception {
-    doTestSortedDocValuesWithSkipper(random().nextInt(1000, 20000));
-  }
-
-  @Nightly
-  public void testSortedDocValuesWithSkipperBig() throws Exception {
-    doTestSortedDocValuesWithSkipper(random().nextInt(50000, 100000));
-  }
-
-  private void doTestSortedDocValuesWithSkipper(int totalDocs) throws Exception {
-    assertDocValuesWithSkipper(
-        totalDocs,
-        new TestDocValueSkipper() {
-          @Override
-          public void populateDoc(Document doc) {
-            doc.add(SortedDocValuesField.indexedField("test", TestUtil.randomBinaryTerm(random())));
-          }
-
-          @Override
-          public DocValuesWrapper docValuesWrapper(LeafReader leafReader) throws IOException {
-            SortedDocValues sortedDocValues = leafReader.getSortedDocValues("test");
-            return new DocValuesWrapper() {
-
-              @Override
-              public int advance(int target) throws IOException {
-                return sortedDocValues.advance(target);
-              }
-
-              @Override
-              public boolean advanceExact(int target) throws IOException {
-                return sortedDocValues.advanceExact(target);
-              }
-
-              @Override
-              public long maxValue() throws IOException {
-                return sortedDocValues.ordValue();
-              }
-
-              @Override
-              public long minValue() throws IOException {
-                return sortedDocValues.ordValue();
-              }
-
-              @Override
-              public int docID() {
-                return sortedDocValues.docID();
-              }
-            };
-          }
-
-          @Override
-          public DocValuesSkipper docValuesSkipper(LeafReader leafReader) throws IOException {
-            return leafReader.getDocValuesSkipper("test");
-          }
-        });
-  }
-
-  public void testSortedSetDocValuesWithSkipperSmall() throws Exception {
-    doTestSortedSetDocValuesWithSkipper(random().nextInt(1, 1000));
-  }
-
-  public void testSortedSetDocValuesWithSkipperMedium() throws Exception {
-    doTestSortedSetDocValuesWithSkipper(random().nextInt(10000, 20000));
-  }
-
-  @Nightly
-  public void testSortedSetDocValuesWithSkipperBig() throws Exception {
-    doTestSortedSetDocValuesWithSkipper(random().nextInt(50000, 100000));
-  }
-
-  private void doTestSortedSetDocValuesWithSkipper(int totalDocs) throws Exception {
-    assertDocValuesWithSkipper(
-        totalDocs,
-        new TestDocValueSkipper() {
-          @Override
-          public void populateDoc(Document doc) {
-            for (int j = 0; j < random().nextInt(1, 5); j++) {
-              doc.add(
-                  SortedSetDocValuesField.indexedField(
-                      "test", TestUtil.randomBinaryTerm(random())));
-            }
-          }
-
-          @Override
-          public DocValuesWrapper docValuesWrapper(LeafReader leafReader) throws IOException {
-            SortedSetDocValues sortedSetDocValues = leafReader.getSortedSetDocValues("test");
-            return new DocValuesWrapper() {
-              long max;
-              long min;
-
-              @Override
-              public int advance(int target) throws IOException {
-                int doc = sortedSetDocValues.advance(target);
-                if (doc != NO_MORE_DOCS) {
-                  readValues();
-                }
-                return doc;
-              }
-
-              @Override
-              public boolean advanceExact(int target) throws IOException {
-                if (sortedSetDocValues.advanceExact(target)) {
-                  readValues();
-                  return true;
-                }
-                return false;
-              }
-
-              private void readValues() throws IOException {
-                max = Long.MIN_VALUE;
-                min = Long.MAX_VALUE;
-                for (int i = 0; i < sortedSetDocValues.docValueCount(); i++) {
-                  long value = sortedSetDocValues.nextOrd();
-                  max = Math.max(max, value);
-                  min = Math.min(min, value);
-                }
-              }
-
-              @Override
-              public long maxValue() {
-                return max;
-              }
-
-              @Override
-              public long minValue() {
-                return min;
-              }
-
-              @Override
-              public int docID() {
-                return sortedSetDocValues.docID();
-              }
-            };
-          }
-
-          @Override
-          public DocValuesSkipper docValuesSkipper(LeafReader leafReader) throws IOException {
-            return leafReader.getDocValuesSkipper("test");
-          }
-        });
-  }
-
-  private void assertDocValuesWithSkipper(int totalDocs, TestDocValueSkipper testDocValueSkipper)
-      throws Exception {
-    Supplier<Boolean> booleanSupplier;
-    switch (random().nextInt(3)) {
-      case 0 -> booleanSupplier = () -> true;
-      case 1 -> booleanSupplier = () -> random().nextBoolean();
-      case 2 -> booleanSupplier = () -> random().nextBoolean() && random().nextBoolean();
-      default -> throw new AssertionError();
-    }
-    Directory directory = newDirectory();
-    RandomIndexWriter writer = new RandomIndexWriter(random(), directory);
-    int numDocs = 0;
-    for (int i = 0; i < totalDocs; i++) {
-      Document doc = new Document();
-      if (booleanSupplier.get()) {
-        testDocValueSkipper.populateDoc(doc);
-        numDocs++;
-      }
-      writer.addDocument(doc);
-      if (rarely()) {
-        writer.commit();
-      }
-    }
-    writer.flush();
-
-    if (random().nextBoolean()) {
-      writer.forceMerge(1);
-    }
-
-    IndexReader r = writer.getReader();
-    int readDocs = 0;
-    for (LeafReaderContext readerContext : r.leaves()) {
-      LeafReader reader = readerContext.reader();
-      ByteArrayOutputStream bos = new ByteArrayOutputStream(1024);
-      PrintStream infoStream = new PrintStream(bos, false, UTF_8);
-      DocValuesStatus status = CheckIndex.testDocValues((CodecReader) reader, infoStream, true);
-      if (status.error != null) {
-        throw new Exception(status.error);
-      }
-      readDocs +=
-          assertDocValuesSkipSequential(
-              testDocValueSkipper.docValuesWrapper(reader),
-              testDocValueSkipper.docValuesSkipper(reader));
-      for (int i = 0; i < 10; i++) {
-        assertDocValuesSkipRandom(
-            testDocValueSkipper.docValuesWrapper(reader),
-            testDocValueSkipper.docValuesSkipper(reader),
-            reader.maxDoc());
-      }
-    }
-    assertEquals(numDocs, readDocs);
-    IOUtils.close(r, writer, directory);
-  }
-
-  private int assertDocValuesSkipSequential(DocValuesWrapper iterator, DocValuesSkipper skipper)
-      throws IOException {
-    if (skipper == null) {
-      return 0;
-    }
-
-    assertEquals(-1, iterator.docID());
-    assertEquals(-1, skipper.minDocID(0));
-    assertEquals(-1, skipper.maxDocID(0));
-
-    iterator.advance(0);
-    int docCount = 0;
-    while (true) {
-      int previousMaxDoc = skipper.maxDocID(0);
-      skipper.advance(previousMaxDoc + 1);
-      assertTrue(skipper.minDocID(0) > previousMaxDoc);
-      if (skipperHasAccurateDocBounds()) {
-        assertEquals(iterator.docID(), skipper.minDocID(0));
-      } else {
-        assertTrue(
-            "Expected: " + iterator.docID() + " but got " + skipper.minDocID(0),
-            skipper.minDocID(0) <= iterator.docID());
-      }
-
-      if (skipper.minDocID(0) == NO_MORE_DOCS) {
-        assertEquals(NO_MORE_DOCS, skipper.maxDocID(0));
-        break;
-      }
-      assertTrue(skipper.docCount(0) > 0);
-
-      int maxDoc = -1;
-      long minVal = Long.MAX_VALUE;
-      long maxVal = Long.MIN_VALUE;
-      for (int i = 0; i < skipper.docCount(0); ++i) {
-        assertNotEquals(NO_MORE_DOCS, iterator.docID());
-        maxDoc = Math.max(maxDoc, iterator.docID());
-        minVal = Math.min(minVal, iterator.minValue());
-        maxVal = Math.max(maxVal, iterator.maxValue());
-        iterator.advance(iterator.docID() + 1);
-      }
-      if (skipperHasAccurateDocBounds()) {
-        assertEquals(maxDoc, skipper.maxDocID(0));
-      } else {
-        assertTrue(
-            "Expected: " + maxDoc + " but got " + skipper.maxDocID(0),
-            skipper.maxDocID(0) >= maxDoc);
-      }
-      if (skipperHasAccurateValueBounds()) {
-        assertEquals(minVal, skipper.minValue(0));
-        assertEquals(maxVal, skipper.maxValue(0));
-      } else {
-        assertTrue(
-            "Expected: " + minVal + " but got " + skipper.minValue(0),
-            minVal >= skipper.minValue(0));
-        assertTrue(
-            "Expected: " + maxVal + " but got " + skipper.maxValue(0),
-            maxVal <= skipper.maxValue(0));
-      }
-      docCount += skipper.docCount(0);
-    }
-
-    assertEquals(docCount, skipper.docCount());
-    return docCount;
-  }
-
-  private static void assertDocValuesSkipRandom(
-      DocValuesWrapper iterator, DocValuesSkipper skipper, int maxDoc) throws IOException {
-    if (skipper == null) {
-      return;
-    }
-    while (true) {
-      int doc = random().nextInt(skipper.maxDocID(0), maxDoc + 1) + 1;
-      skipper.advance(doc);
-      if (skipper.minDocID(0) == NO_MORE_DOCS) {
-        assertEquals(NO_MORE_DOCS, skipper.maxDocID(0));
-        return;
-      }
-      if (iterator.advanceExact(doc)) {
-        assertTrue(iterator.docID() >= skipper.minDocID(0));
-        assertTrue(iterator.docID() <= skipper.maxDocID(0));
-        assertTrue(iterator.minValue() >= skipper.minValue(0));
-        assertTrue(iterator.maxValue() <= skipper.maxValue(0));
-      }
-    }
-  }
-
-  private interface TestDocValueSkipper {
-
-    void populateDoc(Document doc);
-
-    DocValuesWrapper docValuesWrapper(LeafReader leafReader) throws IOException;
-
-    DocValuesSkipper docValuesSkipper(LeafReader leafReader) throws IOException;
-  }
-
-  private interface DocValuesWrapper {
-
-    int advance(int target) throws IOException;
-
-    boolean advanceExact(int target) throws IOException;
-
-    long maxValue() throws IOException;
-
-    long minValue() throws IOException;
-
-    int docID();
   }
 }
